Handle GitHub search limit: dashes in image name
Per #191 and #210, there are a couple of limitations in GitHub search
that effectively limit the ability of DFIU to find code results.

By combining known domain patterns and noting the limitations documented
in https://docs.github.com/en/github/searching-for-information-on-github/searching-code#considerations-for-code-search
this change does some treatment of image names. It treats the TLD and
remaining URL segments slightly different.

The domain will be split into parts by dashes while we'll attempt to
split the remaining url parts by slashes if they have dashes in them.
Meanwhile, we'll attempt to keep the search terms down to a minimum if
possible.

A fair number of working use cases based on manual testing can be
found in the unit tests but working and non-working tests are also
documented in #191 and #210. Some PRs have been created as a result of
this after these changes:
* https://github.com/justinharringa/somecontainers/pull/9
* https://github.com/justinharringa/somecontainers/pull/10
* https://github.com/justinharringa/somecontainers/pull/11
* https://github.com/justinharringa/somecontainers/pull/12
* https://github.com/justinharringa/somecontainers/pull/13

Fixes #210